+++ sdk/containerservice/arm-containerservicefleet/samples/v2-beta/javascript/autoUpgradeProfilesGetSample.js
@@ -8,23 +8,9 @@ const { DefaultAzureCredential } = require("@azure/identity");
  * This sample demonstrates how to get a AutoUpgradeProfile
  *
  * @summary get a AutoUpgradeProfile
- * x-ms-original-file: 2025-08-01-preview/AutoUpgradeProfiles_Get.json
+ * x-ms-original-file: 2026-02-01-preview/AutoUpgradeProfiles_Get.json
  */
 async function getsAnAutoUpgradeProfileResource() {
-  const credential = new DefaultAzureCredential();
-  const subscriptionId = "00000000-0000-0000-0000-000000000000";
-  const client = new ContainerServiceFleetClient(credential, subscriptionId);
-  const result = await client.autoUpgradeProfiles.get("rg1", "fleet1", "autoupgradeprofile1");
-  console.log(result);
-}
-
-/**
- * This sample demonstrates how to get a AutoUpgradeProfile
- *
- * @summary get a AutoUpgradeProfile
- * x-ms-original-file: 2025-08-01-preview/AutoUpgradeProfiles_Get_MaximumSet_Gen.json
- */
-async function getsAnAutoUpgradeProfileResourceGeneratedByMaximumSetRule() {
   const credential = new DefaultAzureCredential();
   const subscriptionId = "00000000-0000-0000-0000-000000000000";
   const client = new ContainerServiceFleetClient(credential, subscriptionId);
@@ -34,7 +20,6 @@ async function getsAnAutoUpgradeProfileResourceGeneratedByMaximumSetRule() {
 
 async function main() {
   await getsAnAutoUpgradeProfileResource();
-  await getsAnAutoUpgradeProfileResourceGeneratedByMaximumSetRule();
 }
 
 main().catch(console.error);